requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pytest==5.3.1
python-dateutil==2.8.1
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.11
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
